#25 Crear componente base
