+++ src/pages/call-for-proposals/components/Technical.js
@@ -1,0 +1,24 @@
+import React from 'react'
+import Container from 'react-bootstrap/Container';
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+
+const Technical = () => {
+  return (
+    <section id="technical">
+      <Container>
+        <Row className="justify-content-center">
+          <Col xs={10} md={8}>
+            We welcome a diverse range of proposals related to various technical topics. Share your
+            expertise on any of the following subjects:
+          </Col>
+          <Col xs={12}>
+
+          </Col>
+        </Row>
+      </Container>
+    </section>
+  );
+}
+
+export default Technical
