+++ src/pages/call-for-proposals/index.js
@@ -2,13 +2,14 @@ import React from 'react'
 
 import Layout from 'components/layout';
 import CallSpeakers from 'components/commons/CallSpeakers';
-import Title from 'components/commons/Title';
+import Title from './components/Title';
+import Content from './components/content';
 
 const callForProposals = () => {
   return (
     <Layout>
       <Title />
-
+      <Content />
       <CallSpeakers />
     </Layout>
   )
