+++ src/static/styles/partials/_call-for-proposals.sass
@@ -1,0 +1,16 @@
+@use "../variables"
+
+.call-for-proposals
+  background-color: variables.$main-bg-color
+
+  #technical
+    border: 1px solid #000
+    background: linear-gradient(153deg, #D3C3ED 0.67%, #D3C3ED 46.7%, #EEE3FF 94.7%)
+    margin: 15px auto
+    padding: 60px 20px
+
+  #topics
+    border: 1px solid #000
+    background: linear-gradient(298deg, #8CB6F3 -7.13%, #8CB6F3 35.52%, #CEE0FB 92.52%)
+    margin: 15px auto
+    padding: 60px 20px
