+++ src/static/styles/styles.sass
@@ -2,3 +2,4 @@
 @use "partials/_index.sass"
 @use "partials/_contact.sass"
 @use "partials/_footer.sass"
+@use "partials/_call-for-proposals.sass"
